[Tizen] Remove the unused code from Application storage
SSIA
